begin branch
The purpose of the modular engine is to make generation of tile transitions easier, allow easier sending of generic chunk data over the network, shrinking ldtk files, and allowing new assets and types to be added quickly during development. This branch breaks and changes a lot of things and may take some significant time to get right.